created base classes and add parsing class - Sentinel
