+++ halo.egg-info/requires.txt
@@ -1,2 +1,3 @@
 pandas
+pandas_profiling
 scikit-learn
