+++ halo/__init__.py
@@ -1,0 +1,1 @@
+from .halo import Covenant, Flood, Sentinel
