--- halo/multimodelingtools/__init__.py
@@ -1,1 +1,0 @@
-from .multimodelingtools import Covenant
